Continuing work on mutation operators for GPTrees.  This required various changes to core EM components such as Population and Genotype, which are now cloneable.  Representations are also cloneable, though without the Cloneable interface due to some odd issue with Java's cloneable not really being a true interface.  GPTree and GPNodes needed to be updated to handle changes and silly things like forgetting to include a backlink to a parent.
Some notes: Populations should not have an assumed structure associated
with them.  Keep add, replace, etc. methods very generic.  Let the
population decide how to handle it by providing parents when swapping or
adding members.  EMStates should be inherited from the clone or, when a
clone is not available, from the first non-null EMState received in a
mutation operator from a downstream population.