Added comments to GPTree material.
Made GPNode a lighter-weight class by using primitives rather than
objects.  This should allow for larger population sizes with greater
tree depths.

Fixed CrossOver.  There was a problem with either linking to a
deprecated parent or ignoring maximum tree depth.  Either way, it was
causing a stack overflow.

Fixed TestPi.  Mutation got deleted at some point.  When adding that
step
back in, errors with crossover appeared.

Added GPTreeUtil to do some common calculations on GPTrees (e.g. find
the maximum depth of a tree or subtree).

Added GPFixedRootTree, which doesn't allow for the root of the tree to
be modified.  This will be useful for the Stimulus/Action system.  I
also had to modify reRoot to handle the new canAlter method, allowing
the tree to test whether or not a node can be altered.

Changed GPNodePos to be lighter weight.  It now uses an int array rather
than an ArrayList of Integers.

Changed GPFullTree to use the correct validation system.

Added validation check to EvolutionPipeline to make sure configurations
are setup correctly.  Modified Pipeline to fit this change.

Added various changes to account for the above.

TestPi works correctly (with new added assumptions about how well it
evolves pi as well as making sure maximum tree height is obeyed).